Remove WASM quickcheck infrastructure
