+++ concordium-std/Cargo.toml
@@ -12,8 +12,6 @@ readme = "./README.md"
 
 [dependencies]
 dlmalloc = { version = "0.2.14", features = ["global"], optional = true }
-quickcheck = {version = "1", optional = true }
-getrandom = { version = "0.2", features = ["custom"], optional = true }
 
 [dependencies.concordium-contracts-common]
 path = "../concordium-rust-sdk/concordium-base/smart-contracts/contracts-common/concordium-contracts-common"
@@ -28,10 +26,8 @@ wasm-test = ["concordium-contracts-common/wasm-test"]
 # Own internal wasm-tests leak out to the smart contracts using this library,
 # so a separate feature 'internal-wasm-test' is introduced for these.
 internal-wasm-test = []
-#internal-wasm-test = ["wasm-test", "concordium-quickcheck"] # todo reenable property based tests as part of https://linear.app/concordium/issue/COR-2474/property-based-tests-on-wasm32-target
 # Make smart contract macros generate function in the smart contract that allows inspecting the schema of the contract functions.
 build-schema = ["concordium-contracts-common/build-schema"]
-concordium-quickcheck = ["getrandom", "quickcheck", "concordium-contracts-common/concordium-quickcheck"] # // decide what to do with this feature flag as part of https://linear.app/concordium/issue/COR-2474/property-based-tests-on-wasm32-target
 # Enabling debug feature changes the #[concoridum_dbg] macro to log messages via the WASM interpreter
 debug = []
 # Use the dlmalloc allocator
